+++ src/Kanbas/Dashboard/index.js
@@ -1,25 +1,12 @@
-import db from "../Database";
 import {Card, Col, Container, Row} from 'react-bootstrap';
 import {Link} from "react-router-dom";
-import {useState} from "react";
 
-function Dashboard() {
-    const [courses, setCourses] = useState(db.courses);
-
-    const [course, setCourse] = useState({
-        name: "New Course",
-        number: "New Number",
-        startDate: "2023-09-10",
-        endDate: "2023-12-15",
-        _id: 0
-    });
-
-    const addNewCourse = () => {
-        // must modify the code such that it would update both course._id as well as the course _id saved in courses
-        const _id = new Date().getTime()
-        updateCourseField("_id", _id)
-        setCourses([...courses, {...course, _id}]);
-    };
+function Dashboard(
+    {
+        courses, course, setCourse, addNewCourse,
+        deleteCourse, updateCourse
+    }
+) {
 
     const updateCourseField = (field, value) => {
         setCourse(prevCourse => ({
@@ -27,22 +14,6 @@ function Dashboard() {
             [field]: value
         }));
     }
-
-    const deleteCourse = (courseId) => {
-        setCourses(courses.filter((course) => course._id !== courseId));
-    };
-
-    const updateCourse = () => {
-        setCourses(
-            courses.map((c) => {
-                if (c._id === course._id) {
-                    return course;
-                } else {
-                    return c;
-                }
-            })
-        );
-    };
 
 
     return (
